moved drops to layer 0, converted keys to ints, primary keys are auto_increment

+++ create_tables_layer_4.sql
@@ -1,16 +1,16 @@
 CREATE TABLE UserDevices
 (
-    UserID      NUMERIC(5),
-    DeviceID    NUMERIC(5),
+    UserID      INT,
+    DeviceID    INT,
     CONSTRAINT  UserDevices_FK1 FOREIGN KEY (UserID)    REFERENCES  Users(UserID),
     CONSTRAINT  UserDevices_FK2 FOREIGN KEY (DeviceID)  REFERENCES  Devices(DeviceID)
 );
 
 CREATE TABLE UserPreferences
 (
-    UserID          NUMERIC(5),
-    PreferenceID    NUMERIC(5),
-    P_Rank          NUMERIC(3),
+    UserID          INT,
+    PreferenceID    INT,
+    P_Rank          INT,
     CONSTRAINT  UserPreferences_FK1 FOREIGN KEY (UserID)        REFERENCES  Users(UserID),
     CONSTRAINT  UserPreferences_FK2 FOREIGN KEY (PreferenceID)  REFERENCES  Preferences(PreferenceID)
 );
